+++ Backend/Dto/ProductDto.cs
@@ -7,7 +7,7 @@ public class ProductDto
     public string? Title { get; set; }
     public required string Description { get; set; }
     public List<string>? Images { get; set; }
-    public required decimal Price { get; set; }
+    public required float Price { get; set; }
     public required DateTime CreatedAt { get; set; }
     public CategoryDto? Category { get; set; }
     public List<TagDto>? Tags { get; set; }
